Add clearer comments, add in-place upgrade support

+++ install_ce.sh
@@ -1,34 +1,52 @@
 #!/bin/bash
+
+# REAL PS3 HANDLING
 if [[ -n "$PS3_IP" ]]; then
+	# Check for RPCS3_PATH_OVERRIDE, which isn't meant to be used with this mode
 	if [[ -n "$RPCS3_PATH_OVERRIDE" ]]; then
 		echo PS3_IP and RPCS3_PATH_OVERRIDE cannot be used together\!
 		exit
 	fi
+
+	# Basically just seeing if the PS3 will connect. If it doesn't, the || is run and the script exits.
 	echo Attempting to connect to PS3...
 	curl --silent "ftp://$PS3_IP/dev_hdd0/game/" || exit
 	echo Connection Established\!
+	
+	# Region checking, the && will only execute if the previous command is successful. This will set a variable that controls the region.
 	curl --silent "ftp://$PS3_IP/dev_hdd0/game/NPEB01162/USRDIR/" && region="NPEB01162" && echo Set European Serial "[$region]".
 	curl --silent "ftp://$PS3_IP/dev_hdd0/game/NPJB00250/USRDIR/" && region="NPJB00250" && echo Set Japanese Serial "[$region]".
 	curl --silent "ftp://$PS3_IP/dev_hdd0/game/NPUB30927/USRDIR/" && region="NPUB30927" && echo Set North American Serial "[$region]".
 	echo Attempting to download rom.psarc from "$region"...
+	
+	# Making a skeleton dev_hdd0 for interoperability with the other parts of the script.
 	mkdir -p fake_dev_hdd0/game/"$region"/USRDIR
 	cd fake_dev_hdd0/game/"$region"/USRDIR || exit
-	curl -O "ftp://$PS3_IP/dev_hdd0/game/$region/USRDIR/rom.psarc" || exit
-	echo Downloaded rom.psarc.
+	
+	# Try to download rom.psarc. If that fails, try to download original.psarc. If that fails, exit the script.
+	curl -O "ftp://$PS3_IP/dev_hdd0/game/$region/USRDIR/rom.psarc" || curl -O "ftp://$PS3_IP/dev_hdd0/game/$region/USRDIR/original.psarc" || exit
+	echo Downloaded your psarc file. If this is a rom.psarc it will be used for installation, if it is an original.psarc it will be used for an upgrade.
+	
 	echo Overriding a few things for later in the script...
+	# A hack to make the later parts of the script work properly.
 	cd ..
 	stfdir=$(pwd)
 fi
 
+# Operating System detected. I plan to implement BSD at some point, since it is UNIX even if very few people use it.
 if [[ "$OSTYPE" == "freebsd"* ]]; then
 	if [[ -n "$RPCS3_PATH_OVERRIDE" ]]; then
 		echo Note: RPCS3_PATH_OVERRIDE unavailable on this platform. Continuing.
 	fi
     echo You must be really brave\!
     echo If you really want this, psarc can probably be compiled from source. Not sure about flips.
+    
+# GNU/Linux handling. This primarily targets Steam Deck and Arch Linux, but will apply to most Linux distributions. Might add musl systems at some point, but honestly if you're running one, you probably don't need help installing patches.
 elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+	# This if statement is just here to skip detecting the game path if you don't actually have it installed.
 	if [[ -n "$PS3_IP" ]]; then
 		echo Continuing with real PS3 setup via FTP...
+	# This handles setting which RPCS3 installation you'd like to use.
 	elif [[ -n "$RPCS3_PATH_OVERRIDE" ]]; then
 		echo Overriding rpcs3 path...
 		if [[ "$RPCS3_PATH_OVERRIDE" == "appimage" ]] || [[ "$RPCS3_PATH_OVERRIDE" == "native" ]]; then
@@ -93,6 +111,8 @@ elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
 	    	exit
 		fi
 	fi
+
+	# Check for game presence and region. It will prioritize the North American version, then the European version, then the Japanese version.
     echo Checking for Sonic The Fighters...
     if [[ -n "$PS3_IP" ]]; then
     	echo StF Directory already set.
@@ -109,27 +129,46 @@ elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
     	echo "Sonic The Fighters not found. Please make sure it is installed with one of the following valid serials: [NPUB30927], [NPEB01162], [NPJB00250]."
         exit
     fi
+
+	# Check to see if we actually have curl installed. If more sanity checks are needed, they will be put here later.
 	echo Running sanity checks...
 	if ! command -v curl &> /dev/null 
 	then
 	    echo "curl not found!"
 	    exit 1
 	fi
+
+	# Beginning actual install process.
 	echo Entering Sonic The Fighters directory...
 	cd "$stfdir/USRDIR" || exit
+	
 	echo Downloading needed files...
 	curl -O https://raw.githubusercontent.com/coatlessali/stf-community-unix/main/patches.tar.gz
 	curl -O https://raw.githubusercontent.com/coatlessali/stf-community-unix/main/flips
 	curl -O https://raw.githubusercontent.com/coatlessali/stf-community-unix/main/psarc
-	
+
+	# Make these executable, otherwise they can't run due to permissions issues.
 	chmod +x ./psarc
 	chmod +x ./flips
-	echo Extracting rom.psarc...
-	./psarc -x rom.psarc
+	
+	# Check for original.psarc and if it's present, do an in-place upgrade rather than an installation.
+	if [ -f original.psarc ]; then
+		echo original.psarc found, assuming in-place upgrade and extracting it...
+		# This will overwrite any patched rom contents with their original equivalent.
+		mv -f original.psarc rom.psarc
+		./psarc -x rom.psarc
+		mv rom.psarc original.psarc
+	# Otherwise, just extract it.
+	else
+		echo Extracting rom.psarc...
+		./psarc -x rom.psarc
+		echo Moving rom.psarc to original.psarc...
+		mv rom.psarc original.psarc
+	fi
+
+	# Cleanup and patching. Most of the echo commands here should serve as decent comments.
 	echo Deleting psarc tool...
 	rm psarc
-	echo Moving rom.psarc to original.psarc...
-	mv rom.psarc original.psarc
 	mv flips patches.tar.gz rom/
 	echo Entering rom directory...
 	cd rom || exit
@@ -153,20 +192,30 @@ elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
 	rm ./*.bps
 	echo Finalizing setup...
 	cd "$stfdir" || exit
+
+	# Download icons and such.
 	echo Downloading Content Information Files...
 	curl https://raw.githubusercontent.com/coatlessali/stf-community-unix/main/cif.tar.gz | tar -xz
+
+	# Skip printing this message if you're doing a real PS3 install over FTP.
 	if [[ -z "$PS3_IP" ]]; then
 		echo Installation Complete\! Have fun\!
 		echo ...and remember: Tux Loves You\!
 		echo " "
 		curl https://raw.githubusercontent.com/coatlessali/stf-community-unix/main/CREDITS.txt
 	fi
+
+# MacOS handling. I wish apples were real.
 elif [[ "$OSTYPE" == "darwin"* ]]; then
     echo MacOS detected!
+
+    # RPCS3_PATH_OVERRIDE only applies to Linux.
     if [[ -n "$RPCS3_PATH_OVERRIDE" ]]; then
     	echo Note: RPCS3_PATH_OVERRIDE unavailable on this platform. Continuing.
     fi
-    # Check for StF
+    
+    # Check for Sonic the Fighters directory.
+    # On MacOS, this will always be stored in the same place, removing some complication.
     gamedir="/Users/$USER/Library/Application Support/rpcs3/dev_hdd0/game"
     if [[ -n "$PS3_IP" ]]; then
     	echo StF Directory already set.
@@ -183,6 +232,8 @@ elif [[ "$OSTYPE" == "darwin"* ]]; then
     	echo Sonic The Fighters not found. Please make sure it is installed with one of the following valid serials: "[NPUB30927], [NPEB01162], [NPJB00250]."
         exit
     fi
+
+	# Check if xcode is installed. We have to compile psarc manually.
     if [ -d $(xcode-select --install) ]; then
         echo xcode installed\!
     else
@@ -190,6 +241,8 @@ elif [[ "$OSTYPE" == "darwin"* ]]; then
        xcode-select --install
        exit
     fi
+
+    # Download and compile psarc.
     echo Downloading psarc tool...
     curl http://ferb.fr/ps3/PSARC/psarc-0.1.3.tar.bz2 -O
     echo Extracting tarball...
@@ -206,13 +259,30 @@ elif [[ "$OSTYPE" == "darwin"* ]]; then
     cd ..
     echo Removing psarc directory...
     rm -rf psarc-0.1.3*
+
     echo Entering "$stfdir/USRDIR/"...
     cd "$stfdir/USRDIR/" || exit
-    ./psarc -x rom.psarc && echo Extracted successfully\!
-	echo Moving rom.psarc to original.psarc...
-	mv rom.psarc original.psarc
+    
+    # Check for original.psarc and if it's present, do an in-place upgrade rather than an installation
+    if [ -f original.psarc ]; then
+    	echo original.psarc found, assuming in-place upgrade and extracting it...
+    	# This will overwrite any patched rom contents with their original equivalent.
+    	mv -f original.psarc rom.psarc
+    	./psarc -x rom.psarc
+    	mv rom.psarc original.psarc
+    # Otherwise, just extract.
+    else
+    	echo Extracting rom.psarc...
+    	./psarc -x rom.psarc
+    	echo Moving rom.psarc to original.psarc...
+    	mv rom.psarc original.psarc
+    fi
+
+	# We don't need this anymore.    
 	echo Removing psarc tool...
 	rm psarc
+
+	# Use curl to download patches, use cmdMultiPatch to apply them. Once again echos should serve as decent comments.
 	echo Entering rom directory...
 	cd rom || exit
 	echo Downloading patches...
@@ -234,8 +304,12 @@ elif [[ "$OSTYPE" == "darwin"* ]]; then
     rm ./*.bps
     rm cmdMultiPatch
     cd "$stfdir" || exit
+
+    # Download icons and whatnot.
     echo Downloading Content Information Files...
     curl https://raw.githubusercontent.com/coatlessali/stf-community-unix/main/cif.tar.gz | tar -xz
+
+    # If real PS3 install, move onto FTPing the files back over. If not, print success and credits.
     if [[ -z "$PS3_IP" ]]; then
     	echo Sonic The Fighters: Community Edition has been installed\!
     	echo " "
@@ -245,18 +319,22 @@ fi
 
 if [[ -n "$PS3_IP" ]]; then
 	cd "$stfdir" || exit
+
+	# Sending icons and whatnot. Curl's ftp implementation is limited, so deleting the originals is necessary.
 	echo Sending Content Information Files to PS3...
 	curl -v "ftp://$PS3_IP/" -Q "DELE dev_hdd0/game/$region/ICON0.PNG"
 	curl -v "ftp://$PS3_IP/" -Q "DELE dev_hdd0/game/$region/PIC0.PNG"
 	curl -T ICON0.PNG "ftp://$PS3_IP/dev_hdd0/game/$region/" --ftp-create-dirs
 	curl -T PIC0.PNG "ftp://$PS3_IP/dev_hdd0/game/$region/" --ftp-create-dirs
+
+	# Send all of the files back upstream.
     cd "$stfdir/USRDIR" || exit
 	echo Moving rom.psarc to original.psarc on PS3...
 	curl -T original.psarc "ftp://$PS3_IP/dev_hdd0/game/$region/USRDIR/"
     curl -v "ftp://$PS3_IP/" -Q "DELE dev_hdd0/game/$region/USRDIR/rom.psarc"
     echo Sending ROM information to PS3...
     cd rom || exit
-    # Good lird
+    # Goodness gracious
     curl -T fontmap.farc "ftp://$PS3_IP/dev_hdd0/game/$region/USRDIR/rom/" --ftp-create-dirs
     curl -T string_array.farc "ftp://$PS3_IP/dev_hdd0/game/$region/USRDIR/rom/" --ftp-create-dirs
     curl -T auth2d/aetdb.bin "ftp://$PS3_IP/dev_hdd0/game/$region/USRDIR/rom/auth2d/" --ftp-create-dirs
@@ -280,5 +358,7 @@ if [[ -n "$PS3_IP" ]]; then
     cd ../../../../../
     echo Installation Complete\!
     echo " "
+
+    # Print credits.
     curl https://raw.githubusercontent.com/coatlessali/stf-community-unix/main/CREDITS.txt
 fi
